requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.6.0
azure-core==1.26.4
azure-identity==1.13.0
cffi==1.15.1
charset-normalizer==2.0.7
chromadb==0.3.22
click==8.0.3
clickhouse-connect==0.5.24
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.95.1
frozenlist==1.3.3
fsspec==2023.5.0
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.14.1
idna==3.3
importlib-metadata==6.6.0
itsdangerous==2.0.1
jc==1.23.1
Jinja2==3.0.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.169
lit==16.0.3
llama-index==0.6.8
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.0.9
packaging==23.1
pandas==2.0.1
pandas-stubs==1.2.0.35
Pillow==9.5.0
pipdeptree==2.7.1
portalocker==2.7.0
posthog==3.0.1
protobuf==3.20.3
pyarrow==12.0.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2023.5.5
requests==2.29.0
rich==13.3.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
starlette==0.26.1
streamlit==1.24.0
streamlit-chat==0.0.2.2
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tqdm==4.62.3
transformers==4.29.1
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.7
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
websockets==11.0.3
Werkzeug==2.0.2
xmltodict==0.13.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
tornado==6.3.2
blinker==1.6.2
cachetools==5.3.0
certifi==2021.10.8
watchdog==3.0.0